split TRN into video_trn & TA3N to a separate class
